refactor(portfolio-contract): move planner library to tools/
