+++ packages/portfolio-contract/tools/network/path.ts
@@ -1,6 +1,9 @@
 import type { Amount } from '@agoric/ertp';
 import type { RebalanceGraph, FlowEdge } from '../plan-solve.js';
-import type { AssetPlaceRef, MovementDesc } from '../type-guards-steps.js';
+import type {
+  AssetPlaceRef,
+  MovementDesc,
+} from '../../src/type-guards-steps.js';
 
 /** Weight selector: cheapest => variableFee, fastest => timeFixed, default 1 */
 const edgeWeight = (e: FlowEdge, mode: 'cheapest' | 'fastest') => {
